requirements.txt
selenium==4.3.0
pytest==7.1.2
pytest-bdd==6.0.1
pytest-xdist==2.5.0
webdriver-manager==3.8.2
allure-pytest==2.9.45